added select council in program addition

+++ app/Livewire/ManageProgram/DeleteProgram.php
@@ -2,12 +2,16 @@
 
 namespace App\Livewire\ManageProgram;
 
+use App\Events\UserActionUpdated;
 use App\Models\Program;
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Validation\ValidationException;
 use Livewire\Component;
 
 class DeleteProgram extends Component
 {
     public $program;
+    public $password;
 
     public function mount($programId)
     {
@@ -16,6 +20,25 @@ class DeleteProgram extends Component
     }
     public function delete()
     {
+        $this->validate([
+            'password' => 'required|string',
+        ]);
+
+
+        // Verify the provided password matches the user's password
+        if (!Hash::check($this->password, auth()->user()->password)) {
+            throw ValidationException::withMessages([
+                'password' => 'The password does not match our records.',
+            ]);
+        }
+
+        if (!$this->password) {
+            throw ValidationException::withMessages([
+                'password' => 'The password cannot be empty.',
+            ]);
+        }
+        $this->password = '';
+
         if ($this->program) {
             $this->program->delete();
             $this->dispatch('program-deleted');
